Deducation/requirements.txt
django-filter
django
drf-yasg
coreapi
djangorestframework_simplejwt